Extract lando extras class to own file
